moved db_handler initialization into new file to be shared by any api.  Also corrected import * in tenant/resources.py
